requirements.txt
Django==5.1.1
requests==2.31.0
python-dotenv==1.0.0
django-allauth==0.61.0
djangorestframework==3.15.0
django-cors-headers==4.3.1
python-dateutil==2.8.2
django-chartjs==2.3.0
gunicorn==21.2.0
whitenoise==6.6.0 
